+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 
 Big things planned! More unit tests! Less ORM quirks! More awesomeness!
 
+### Features
+
+* **crud-controller:** add transformInclude hook as a temporary replacement for scopes
+
 
 
 <a name="0.4.0"></a>